move study area calculation to tools, expand study areas to match datasets
